[export] Add stack_trace metadata for non-strict, remove from placeholder/output nodes, add checks to verifier (#121034)
Summary:
This PR addresses 2 issues with stack_trace metadata:
- stack_trace is currently missing from nodes in non-strict export
- in strict mode, stack_trace is populated for placeholder nodes, which may not be well-defined (with multiple uses)

The PR implements:
- adds an extra check to populate stack_trace for nodes, through torch/fx/proxy.py
- removes stack_trace for placeholder/output nodes
- adds a node-level check to the Verifier class

Note: the verifier checks that stack_trace is None for placeholder/output nodes, and None **or** a non-empty str for call_function/get_attr nodes. Ideally for the latter we should just be able to assert that they're non-empty strs, (the result of torch.export.export() should comply with this), but intermediate ExportedProgram objects might not comply with this (e.g. nodes generated for inline range constraints do not have stack_trace populated)


Differential Revision: D54917080

Pulled By: pianpwk