Add torch.library.def_blackbox
This is the entrypoint for defining a blackbox (e.g. PyTorch will
never peek into it) custom op. In this PR, you can specify backend impls
and the abstract impl for this op.

NB: most of this PR is docstrings, please don't be intimidated by the
line count.

There are a number of interesting features:
- we infer the schema from type hints. In a followup I add the ability
  to manually specify a schema.
- we infer the namespace for the operator from the module. In a
  followup, we'll add the ability to manually specify a namespace.
- def_blackbox registrations can override each other. This makes them
  more pleasant to work with in environments like colab.
- we require that the outputs of the blackbox op do not alias any inputs
  or each other. We enforce this via a runtime check.

Test Plan:
- new tests

ghstack-source-id: 165cd88503aff64b92898691823b2c89bfdf0a39
Pull Request resolved: https://github.com/pytorch/pytorch/pull/122344